refactor: extract subagent session handlers
