+++ package.json
@@ -39,6 +39,7 @@
     "src/multiplexer-tmux.ts",
     "src/multiplexer-zellij.ts",
     "src/rehydrate.ts",
+    "src/session-handlers.ts",
     "src/subagent.ts",
     "src/subagent-artifact-cli.ts",
     "src/workflow.ts"
